controller에 input, output view 분리
